pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-materials"
version = "0.1.0"
description = "RL-driven materials discovery using Quantum ESPRESSO"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "ase>=3.22",
    "torch>=2.0",
    "gymnasium>=0.29",
    "stable-baselines3>=2.0",
    "tensorboard>=2.14",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pymatgen>=2023.7",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "jarvis-tools>=2023.7",
    "mp-api>=0.40",
    "requests>=2.28",
]

[project.optional-dependencies]
gnn = [
    "torch-geometric>=2.3",
    "e3nn>=0.5",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.7",
    "ruff>=0.0.280",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
